fix: [#S10P12A402-98] edit Long type to Integer type

+++ backend/src/main/java/org/ssafy/pasila/domain/live/entity/Live.java
@@ -45,7 +45,7 @@ public class Live {
     private String fullVideoUrl;
 
     @Column(name = "like_cnt")
-    private Long likeCnt;
+    private Integer likeCnt;
 
     @Column(name = "is_on")
     private boolean isOn;
